corregir migraciones de base de datos, definicion de variables y rutas de archivos SQL

+++ database/migrations/20260521_155659_insert_nivel_acceso.go
@@ -19,12 +19,15 @@ func init() {
 
 // Run the migrations
 func (m *InsertNivelAcceso_20260521_155659) Up() {
-	m.SQL("INSERT INTO sistema.nivel_acceso (nombre, fecha_asignacion) VALUES ('super admin', '2026-05-20 17:14:45.839255');")
-
+	m.SQL(`INSERT INTO sistema.nivel_acceso (nombre, fecha_asignacion) VALUES 
+		('super admin', '2026-05-20 17:14:45.839255'),
+		('admin', '2026-05-20 17:14:45.839255'),
+		('moderador', '2026-05-20 17:14:45.839255'),
+		('soporte', '2026-05-20 17:14:45.839255'),
+		('auditor', '2026-05-20 17:14:45.839255');`)
 }
 
 // Reverse the migrations
 func (m *InsertNivelAcceso_20260521_155659) Down() {
-	m.SQL("DELETE FROM sistema.nivel_acceso WHERE nombre = 'super admin' ")
-
+	m.SQL("DELETE FROM sistema.nivel_acceso WHERE nombre IN ('super admin', 'admin', 'moderador', 'soporte', 'auditor');")
 }
